Add GitHub Workflows support for some basic CI jobs #5
### Summary

Adds a GH workflow for some basic CI tests, specifically it adds:

    cargo check - does the lib and all the examples compile okay
    cargo test - runs tests
    cargo clippy - linter for code-quality

Since things are still very much in the prototype'y phase, the linter
is set to allow warnings. Does clippy emit higher values than that?
Anyways, once things become more mature, we can circle back to turning
this on and making it more aggressive (and probably a giant
refactor/cleanup when that happens)

### Motivation

CI is a good thing? A nice status badge is cool.

### Test Plan

This is the test plan.