add cache (#24)
### Summary

Add per-job caching to CI using swatinem/ruts-cache.

### Motivation

  + Faster builds (and hopefully less load on github CI)
  + Resolves #7 

### Test Plan

  + CI should pass as normal
  + Manually re-ran CI jobs to observe successful cache behavior

Don't have a great way of seeing how this will do across multiple
branches and, being for a library, may have limited effectiveness. So
this will have to be something I observe over time and ensure it's
behaving in an effective way.