Impleent `ActorAddress` resolution
Added a method to the `RuntimeManagerRef` that can communicate with the
management thread in a pseudo-blocking mechanism (only blocks the
calling thread) in order to perform address resolution.

This should be sufficient once implementing message sending to locate
the right place to send a message and cache the `Sender` in the
`ActorAddress`.