Unsupported message type handler (#48)
### Summary

Adds and `unhandled` method to the `Actor` trait. The method should be
called by the Actor when it is unable to process an incoming message.
The new method currently emits warning log messages, but will eventually
be wried up to the dead-letter queue.

### Motivation

Handling unexpected/unknown messages is a core part of the API and this
satisfies that API surface while leaving the implementation as mostly a
TODO.

### Test Plan

Compilation and usage in the example is currently sufficient. This may
no longer be sufficient once the dead-letter queue system is built.