[wip] Implement sender (#50)
### Summary

Adds a `sender()` type to the `Context` object to allow for actors to
respond to the sender of a message. Also updates the ping-pong example
to make use of this to run a continuous back-and-forth example.

+ Add `Letter` type for internal representation of messages
+ Add `Uri` type to formalize addresses in ActorAddress
+ Fixes bug in execution loop of `ThreadExecutor`
+ Adds `sender()` to `Context`
+ Adds `parent` and `children` to `AcotrCell`

### Motivation

Receiving the sender of a message maybe isn't _critical_, but it's a big
enough ergonomic hurdle that not having it is a realistic blocker for
most people. Even in the ping-pong example, this would require sending a
separate message to share the address before sending the first ping.

#11 

### Test Plan

Things are still early in the project so like most other PRs that have
landed, simply having CI compile, lint, and run existing tests is
_mostly_ sufficient. But in addition to that, this change will:

+ [x] Update the ping-pong example to actually ping and pong
continuously
+ [x] Add unit-test coverage for the new `Uri` type
+ [x] Add unit-test coverage for `ActorAddress` for the bits related to
`Uri`