Update documentation and licensing info in README.md (#64)
### Summary

Rephrase the documentation section in the README now that the docs.rs
stuff is semi-useful. Also add some bits on how to run the examples
since I think that's going to be a decent place to look at code going
forward and I expect to add more examples.

Add a section regarding the license at the bottom. I'm not sure if the
GPLv3 is super popular and folks may not expect it, so worth pointing
out.

### Motivation

README.md was lagging a bit behind the project.

### Test Plan

Spell check in CI