add merge2 to busan::message::Message (#65)
### Summary

Add `merge2` to `Message` which is a wrapper to `prost::Message::merge`.
Added as an undocumented (hidden) method on Message that is primarily an
implementation detail internal to busan.

### Motivation

The `Message::merge` function from prost requires that self be `Sized`.
`merge2` allows us to call the merge function through dynamic dispatch
by implementing `merge2` on the object and not in the trait.

### Test Plan

It compiles. Usages will come later for further testing/refinement.