Split message sending into two (complementary) methods (#66)
### Summary

Split message sending into two methods:

+ `Context::send` - Utility method that takes `ToMessage`, performs
conversion and calls `Context::send_message`
+ `Context::send_message` - Takes a `Box<Message>`, resolves the
address, performs serialization/deserialization (in debug mode), and
transmits the message.

### Motivation

The previous implementation of `send_message` worked by using
`ToMessage` as a convenience type to cut down on the boilerplate of
sending messages. This worked great, however it had one drawback in that
is relied on concrete types.

The ability to also send a `Box<Message>` allows for patterns,
abstractions, or middle-men to be built. Take a load-balancer actor for
example. It receives `Box<Message>`s and then forwards them to an actor
in a pool. In this instance, there is no concrete type we're dealing
with so it is difficult to use `ToMessage`.

### Test Plan

The current examples and tests should cover the new use-case as the new
method is still in the critical path.