support sending ActorAddress objects (#67)
### Summary

More or less what the title says. This PR adds the proto messages and
necessary trait implementations to send an `ActorAddress` (or a slice of
them) as a `busan::message::Message`.

In theory this means they can be serialized and shared across process
boundaries, but in reality that would require a bit more work as nothing
other than the "local" scheme is supported at the moment.

### Motivation

Sharing addresses unlocks more dynamic messaging patterns as currently
it's only possible to send messages between parents and children.

### Test Plan

Nothing ATM. Work on load balancing will make use of and test this
further.