[actor shutdown] Add a context macro to thread_executor (#72)
### Summary

Introduces a new `context!` macro in thread_executor that just makes
constructing context objects less visually intrusive (8 lines per
instantiation!). Code is much easier to read now.

### Motivation

Part of #29.

The changes for actor shutdown will introduce more instances in which a
context object will be created from a system context. It made sense to
invest in a macro here to save some typing. Took a few iterations to get
this working right (so many TIL moments).

### Test Plan

If it compiles it's good. This change should be functionally equivalent
to what existed before.