[thread_executor] Fix thread sleep in work loop (#73)
### Summary

Fix the loop in the thread executor to only sleep when no work has been
done on an iteration. Because only one message per actor is processed
per loop (for fairness or whatever), it's very possible/likely for there
to be more message waiting. In those instances, we don't need to sleep.

### Motivation

Found this issue while working on #29. We shouldn't sleep when there is
still work to be done (potentially). Even though the thread-executor
will eventually be replaced with a tokio runtime, should still fix this
up.

### Test Plan

Existing examples thoroughly exercise this.