[actor shutdown] shutdown children (#75)
### Summary

Propegates shutdown signals to children via the runtime manager.
Currently there is no coordination between timing/ordering of parents
and children except for the flow of initial messages. That is to say
that the actual moment in which actors are removed from the runtime
queue and the `before`/`after_stop` methods are finished may race. I
think this is a fine trade-off for now. Coordination is a lot trickier,
and possibly not necessary.

### Motivation

Part of #29.

Actors are organized in heirarchical trees. In an actor system, all
actors should be part of the same tree and thus if an actor shuts down
and has children, the shutdown signals have to propegate down the tree.

### Test Plan

TBD. Will address testing as the final step in impelemnting actor
shutdown.